Create dataset
Create a directory for image storage

Added a new directory named "datadet" to the project repository. This directory will be used to store images processed by the code. Having a dedicated directory for image storage helps organize the project and ensures easy access to the images

This commit creates the "dataset" directory in the repository.